# artifacts/api-server/src/routes/schools.ts
import { Router } from "express";
import { db, schoolsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { requireAuth } from "../lib/auth";
import { z } from "zod";

const router = Router();

function slugify(name: string): string {
  return name.toLowerCase().replace(/[^a-z0-9]+/g, "-").replace(/(^-|-$)/g, "");
}

const createSchoolSchema = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  phone: z.string().optional(),
  address: z.string().optional(),
  country: z.string().default("Ghana"),
  plan: z.enum(["free", "basic", "pro", "enterprise"]).default("free"),
});

const updateSchoolSchema = z.object({
  name: z.string().optional(),
  email: z.string().email().optional(),
  phone: z.string().optional(),
  address: z.string().optional(),
  logoUrl: z.string().optional(),
  isActive: z.boolean().optional(),
  plan: z.enum(["free", "basic", "pro", "enterprise"]).optional(),
});

router.get("/", requireAuth, async (req, res) => {
  const schools = await db.select().from(schoolsTable).orderBy(schoolsTable.createdAt);
  res.json(schools);
});

router.post("/", requireAuth, async (req, res) => {
  try {
    const data = createSchoolSchema.parse(req.body);
    const slug = slugify(data.name) + "-" + Date.now().toString(36);
    const [school] = await db.insert(schoolsTable).values({
      ...data,
      slug,
    }).returning();
    res.status(201).json(school);
  } catch (err: any) {
    if (err?.name === "ZodError") {
      res.status(400).json({ error: "Validation Error", message: err.errors[0]?.message || "Invalid data" });
    } else {
      res.status(500).json({ error: "Internal Error", message: "Failed to create school" });
    }
  }
});

router.get("/:schoolId", requireAuth, async (req, res) => {
  const [school] = await db.select().from(schoolsTable).where(eq(schoolsTable.id, req.params.schoolId)).limit(1);
  if (!school) {
    res.status(404).json({ error: "Not Found", message: "School not found" });
    return;
  }
  res.json(school);
});

router.patch("/:schoolId", requireAuth, async (req, res) => {
  try {
    const data = updateSchoolSchema.parse(req.body);
    const [school] = await db.update(schoolsTable)
      .set({ ...data, updatedAt: new Date() })
      .where(eq(schoolsTable.id, req.params.schoolId))
      .returning();
    if (!school) {
      res.status(404).json({ error: "Not Found", message: "School not found" });
      return;
    }
    res.json(school);
  } catch (err: any) {
    res.status(500).json({ error: "Internal Error", message: "Failed to update school" });
  }
});

export default router;