# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: BallotWave - Digital Voting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a new school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{schoolId}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get school by ID
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: School details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    patch:
      operationId: updateSchool
      tags: [schools]
      summary: Update school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchoolRequest"
      responses:
        "200":
          description: School updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

  # Elections
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/verify-otp:
    post:
      operationId: verifyOtp
      tags: [users]
      summary: Verify OTP for voter identity
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/send-otp:
    post:
      operationId: sendOtp
      tags: [users]
      summary: Send OTP to phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [super_admin, school_admin, candidate, voter]
        schoolId:
          type: string
        studentId:
          type: string
        isVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, isVerified, isPhoneVerified, createdAt]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [super_admin, school_admin, candidate, voter]
    SendOtpRequest:
      type: object
      properties:
        phone:
          type: string
      required: [phone]
    VerifyOtpRequest:
      type: object
      properties:
        otp:
          type: string
      required: [otp]

    # School
    School:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        country:
          type: string
        logoUrl:
          type: string
        plan:
          type: string
          enum: [free, basic, pro, enterprise]
        isActive:
          type: boolean
        totalElections:
          type: integer
        totalVoters:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, plan, isActive, totalElections, totalVoters, createdAt]
    CreateSchoolRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        country:
          type: string
        plan:
          type: string
          enum: [free, basic, pro, enterprise]
      required: [name, email, country]
    UpdateSchoolRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        logoUrl:
          type: string
        isActive:
          type: boolean
        plan:
          type: string
          enum: [free, basic, pro, enterprise]

    # Elections
...